Prevent exception “The process has already exited” when trying to kill the bot process.

+++ ChallengeHarness/Runners/BotRunner.cs
@@ -194,7 +194,8 @@ namespace ChallengeHarness.Runners
 
             if (!didExit)
             {
-                p.Kill();
+                if (!p.HasExited)
+                    p.Kill();
                 OutputAppendLog(String.Format("[GAME]\tBot {0} timed out after {1} ms.", PlayerName,
                     _botTimer.ElapsedMilliseconds));
                 OutputAppendLog(String.Format("[GAME]\tKilled process {0}.", _processName));
